refactor: move new task button to navbar
